Modified methods to use int64. This will still cause errors as blindly changing int32_t to int64_t affects more tests, including small_file test that expects 4-byte headers. More work needs to be done to fix these things.

+++ chunked_ostream.cc
@@ -53,7 +53,7 @@ namespace libdap {
 std::streambuf::int_type chunked_outbuf::data_chunk() {
     DBG(cerr << "In chunked_outbuf::data_chunk" << endl);
 
-    int32_t num = pptr() - pbase(); // num needs to be signed for the call to pbump
+    int64_t num = pptr() - pbase(); // num needs to be signed for the call to pbump
 
     // Since this is called by sync() (e.g., flush()), return 0 and do nothing
     // when there's no data to send.
@@ -71,7 +71,7 @@ std::streambuf::int_type chunked_outbuf::data_chunk() {
     // much anyway... Here's the general idea all the same:
     //
     // unsigned int chunk_header = (unsigned int)num | CHUNK_type;
-    uint32_t header = num;
+    uint64_t header = num;
 
     // Add encoding of host's byte order. jhrg 11/24/13
     if (!d_big_endian)
@@ -106,13 +106,13 @@ std::streambuf::int_type chunked_outbuf::data_chunk() {
 std::streambuf::int_type chunked_outbuf::end_chunk() {
     DBG(cerr << "In chunked_outbuf::end_chunk" << endl);
 
-    int32_t num = pptr() - pbase(); // num needs to be signed for the call to pbump
+    int64_t num = pptr() - pbase(); // num needs to be signed for the call to pbump
 
     // write out the chunk headers: CHUNKTYPE and CHUNKSIZE
     // as a 32-bit unsigned int. Here I assume that num is never
     // more than 2^24 because that was tested in the constructor
 
-    uint32_t header = (uint32_t)num | CHUNK_END;
+    uint64_t header = (uint64_t)num | CHUNK_END;
 
     // Add encoding of host's byte order. jhrg 11/24/13
     if (!d_big_endian)
@@ -149,7 +149,7 @@ std::streambuf::int_type chunked_outbuf::err_chunk(const std::string &m) {
 
     // Figure out how many chars are in the buffer - these will be
     // ignored.
-    int32_t num = pptr() - pbase(); // num needs to be signed for the call to pbump
+    int64_t num = pptr() - pbase(); // num needs to be signed for the call to pbump
 
     // write out the chunk headers: CHUNKTYPE and CHUNKSIZE
     // as a 32-bit unsigned int. Here I assume that num is never
@@ -157,7 +157,7 @@ std::streambuf::int_type chunked_outbuf::err_chunk(const std::string &m) {
     if (msg.length() > 0x00FFFFFF)
         msg = "Error message too long";
 
-    uint32_t header = (uint32_t)msg.length() | CHUNK_ERR;
+    uint64_t header = (uint64_t)msg.length() | CHUNK_ERR;
 
     // Add encoding of host's byte order. jhrg 11/24/13
     if (!d_big_endian)
@@ -247,7 +247,7 @@ std::streamsize chunked_outbuf::xsputn(const char *s, std::streamsize num) {
     // the bytes remaining 's' in the buffer. Return the number of bytes sent
     // or 0 if an error is encountered.
 
-    int32_t bytes_in_buffer = pptr() - pbase(); // num needs to be signed for the call to pbump
+    int64_t bytes_in_buffer = pptr() - pbase(); // num needs to be signed for the call to pbump
 
     // Will num bytes fit in the buffer? The location of epptr() is one back from
     // the actual end of the buffer, so the next char written will trigger a write
@@ -261,7 +261,7 @@ std::streamsize chunked_outbuf::xsputn(const char *s, std::streamsize num) {
 
     // If here, write a chunk header and a chunk's worth of data by combining the
     // data in the buffer and some data from 's'.
-    uint32_t header = d_buf_size;
+    uint64_t header = d_buf_size;
 
     // Add encoding of host's byte order. jhrg 11/24/13
     if (!d_big_endian)
@@ -270,7 +270,7 @@ std::streamsize chunked_outbuf::xsputn(const char *s, std::streamsize num) {
     // network byte order for the header
     header = htonl(header);
 
-    d_os.write((const char *)&header, sizeof(int32_t)); // Data chunk's CHUNK_TYPE is 0x00000000
+    d_os.write((const char *)&header, sizeof(int64_t)); // Data chunk's CHUNK_TYPE is 0x00000000
 
     // Reset the pptr() and epptr() now in case of an error exit. See the 'if'
     // at the end of this for the only code from here down that will modify the
@@ -286,13 +286,13 @@ std::streamsize chunked_outbuf::xsputn(const char *s, std::streamsize num) {
     if (d_os.eof() || d_os.bad())
         return traits_type::not_eof(0);
     s += bytes_to_fill_out_buffer;
-    uint32_t bytes_still_to_send = num - bytes_to_fill_out_buffer;
+    uint64_t bytes_still_to_send = num - bytes_to_fill_out_buffer;
 
     // Now send all the remaining data in s until the amount remaining doesn't
     // fill a complete chunk and buffer those data.
     while (bytes_still_to_send >= d_buf_size) {
         // This is header for  a chunk of d_buf_size bytes; the size was set above
-        d_os.write((const char *)&header, sizeof(int32_t));
+        d_os.write((const char *)&header, sizeof(int64_t));
         d_os.write(s, d_buf_size);
         if (d_os.eof() || d_os.bad())
             return traits_type::not_eof(0);
